+++ apps/journeys-admin/src/components/JourneyList/JourneyCard/JourneyCardInfo/index.ts
@@ -1,0 +1,1 @@
+export { JourneyCardInfo } from './JourneyCardInfo'
