+++ apps/journeys-admin/src/components/JourneyList/JourneyCard/JourneyCardText/index.ts
@@ -1,0 +1,1 @@
+export { JourneyCardText } from './JourneyCardText'
